Refactor consensus code for better modularity and maintainability
Refactor the ICN consensus codebase to improve modularity and maintainability.

* **Configuration Refactoring**
  - Move `ConsensusConfig`, `ValidatorConfig`, `RoundConfig`, `EventConfig`, and `MetricsConfig` structs and their implementations from `crates/icn-consensus/src/config.rs` to separate files in the `config` directory.
  - Remove the moved structs and implementations from `crates/icn-consensus/src/config.rs`.
  - Add `mod` statements in `crates/icn-consensus/src/config.rs` to import the moved modules.

* **Cryptographic Operations Refactoring**
  - Move `sign`, `verify`, and `hash_message` functions from `crates/icn-consensus/src/crypto.rs` to `crates/icn-consensus/src/crypto/signing.rs`.
  - Move `generate_keypair` function from `crates/icn-consensus/src/crypto.rs` to `crates/icn-consensus/src/crypto/keypair.rs`.
  - Remove the moved functions from `crates/icn-consensus/src/crypto.rs`.
  - Add `mod` statements in `crates/icn-consensus/src/crypto.rs` to import the moved modules.

* **Error Handling Refactoring**
  - Move `ConsensusError` enum from `crates/icn-consensus/src/error.rs` to `crates/icn-consensus/src/error/consensus_error.rs`.
  - Remove the moved enum from `crates/icn-consensus/src/error.rs`.
  - Add `mod` statement in `crates/icn-consensus/src/error.rs` to import the moved module.

* **Proof of Cooperation Refactoring**
  - Break down methods in `ProofOfCooperation` struct in `crates/icn-consensus/src/proof_of_cooperation/core.rs` into smaller, more focused functions.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/InterCooperative-Network/ICN/tree/feature/workspace-reorganization?shareId=XXXX-XXXX-XXXX-XXXX).
