+++ crates/icn-consensus/src/crypto/signing.rs
@@ -1,10 +1,6 @@
 use secp256k1::{Secp256k1, Message, SecretKey, PublicKey, Signature};
 use sha2::{Sha256, Digest};
 use thiserror::Error;
-use crate::error::{ConsensusError, ConsensusResult};
-
-mod signing;
-mod keypair;
 
 #[derive(Error, Debug)]
 pub enum CryptoError {
@@ -30,21 +26,29 @@ impl CryptoManager {
         }
     }
 
-    /// Integrate cryptographic operations with the consensus engine
-    pub fn integrate_with_consensus(&self, consensus: &mut crate::ConsensusEngine) -> ConsensusResult<()> {
-        // Example integration logic
-        let (secret_key, public_key) = self.generate_keypair()?;
-        let message = b"consensus message";
-        let signature = self.sign(message, &secret_key)?;
-        let is_valid = self.verify(message, &signature, &public_key)?;
+    /// Sign a message with a secret key
+    pub fn sign(&self, message: &[u8], secret_key: &SecretKey) -> CryptoResult<Signature> {
+        let message = self.hash_message(message);
+        let message = Message::from_slice(&message)
+            .map_err(|e| CryptoError::SigningError(e.to_string()))?;
+        
+        Ok(self.secp.sign_ecdsa(&message, secret_key))
+    }
 
-        if is_valid {
-            consensus.process_signature(public_key, signature)?;
-        } else {
-            return Err(ConsensusError::InvalidSignature);
-        }
+    /// Verify a signature
+    pub fn verify(&self, message: &[u8], signature: &Signature, public_key: &PublicKey) -> CryptoResult<bool> {
+        let message = self.hash_message(message);
+        let message = Message::from_slice(&message)
+            .map_err(|e| CryptoError::SigningError(e.to_string()))?;
+        
+        Ok(self.secp.verify_ecdsa(&message, signature, public_key).is_ok())
+    }
 
-        Ok(())
+    /// Hash a message using SHA256
+    fn hash_message(&self, message: &[u8]) -> [u8; 32] {
+        let mut hasher = Sha256::new();
+        hasher.update(message);
+        hasher.finalize().into()
     }
 }
